cleanup: move plot. strat. to bench file
