+++ examples/sat/bench_hillclimber.py
@@ -7,13 +7,13 @@ import logging
 from datetime import datetime
 from os import path, makedirs
 from pathlib import Path
-
 import click
 import numpy as np
 import pandas as pd
 
+from qubrabench.utils.plotting_strategy import PlottingStrategy
+
 import hillclimber
-from hillclimber_plotting_strategy import HillClimberPlottingStrategy
 
 
 @click.group()
@@ -127,6 +127,45 @@ def setup_default_logger(verbose: bool):
         )
         file_handler.setFormatter(log_formatter)
         root_logger.addHandler(file_handler)
+
+
+class HillClimberPlottingStrategy(PlottingStrategy):
+    """
+    Plot data generated by hill-climb vs Cade et al.
+    Generate one plot for each (k, r) pair.
+    """
+
+    def __init__(self):
+        self.colors["QuBRA"] = "blue"
+        self.colors["Cade"] = "orange"
+
+    def columns_to_group_for_plots(self):
+        return ["k", "r"]
+
+    def columns_to_group_in_a_plot(self):
+        return ["impl"]
+
+    def compute_aggregates(self, data, *, quantum_factor):
+        # compute combined query costs of quantum search
+        c = data["quantum_expected_classical_queries"]
+        q = data["quantum_expected_quantum_queries"]
+        data["quantum_cost"] = c + quantum_factor * q
+        return data
+
+    def x_axis_column(self):
+        return "n"
+
+    def x_axis_label(self):
+        return "$n$"
+
+    def y_axis_label(self):
+        return "Queries"
+
+    def columns_to_plot(self):
+        return {
+            "classical_actual_queries": ("Classical Queries", "o"),
+            "quantum_cost": ("Quantum Queries", "x"),
+        }
 
 
 @cli.command()
