+++ qubrabench/utils/plotting_strategy.py
@@ -18,6 +18,69 @@ class PlottingStrategy(ABC):
     # colors for each dataset in the plot
     colors: dict[str, str] = {}
 
+    @abstractmethod
+    def x_axis_column(self) -> str:
+        """
+        Data column to plot along the X-axis
+        """
+        return ""
+
+    @abstractmethod
+    def x_axis_label(self) -> str:
+        """
+        Label to display for the X-axis
+        """
+        return ""
+
+    @abstractmethod
+    def y_axis_label(self) -> str:
+        """
+        Label to display for the Y-axis
+        """
+        return ""
+
+    @abstractmethod
+    def columns_to_group_for_plots(self) -> list[str]:
+        """
+        TODO better name
+
+        Generate a plot for each unique tuple of values for the specified columns.
+        Example: ["k", "n"] - a plot will be generated for each unique tuple value (k, n).
+        Example: [] - generate a single plot with the entire data.
+
+        Returns:
+            List of column names to group by.
+        """
+        return []
+
+    @abstractmethod
+    def columns_to_group_in_a_plot(self) -> list[str]:
+        """
+        TODO better name
+
+        Generate a data line for each unique value in the specified columns.
+
+        Example: ["impl"] - a line will be generated for each unique `impl` label.
+        """
+
+    @abstractmethod
+    def compute_aggregates(
+        self, data: pd.DataFrame, *, quantum_factor: float
+    ) -> pd.DataFrame:
+        return data
+
+    @abstractmethod
+    def columns_to_plot(self) -> dict[str, tuple[str, str]]:
+        """
+        Dictionary of columns to display in the plot.
+            Key: is the column name in the dataframe
+            Value: Column display name, Marker to use in the plot
+
+        Example:
+            {"c": ("Classical", "o"), "q": ("Quantum", "x")}
+        """
+        return {}
+
     def plot(self, data: pd.DataFrame, *, quantum_factor: float = 2):
         """
         Plot benchmarking data.
@@ -52,14 +115,14 @@ class PlottingStrategy(ABC):
             ax.set_ylim(300, 10**y_scale_exponent)
             ax.set_xscale("log")
             ax.set_yscale("log")
-            ax.set_xlabel(self.get_x_label())
-            ax.set_ylabel(self.get_y_label())
+            ax.set_xlabel(self.x_axis_label())
+            ax.set_ylabel(self.y_axis_label())
             ax.grid(which="both")
 
             # group lines by implementation
             impls = group.groupby(self.columns_to_group_in_a_plot())
             for impl_params, impl in impls:
-                plot_data = impl.groupby(self.get_x_axis_column())
+                plot_data = impl.groupby(self.x_axis_column())
                 means = plot_data.mean(numeric_only=True)
                 errors = plot_data.sem(numeric_only=True)
 
@@ -90,63 +153,7 @@ class PlottingStrategy(ABC):
         plt.subplots_adjust(top=0.7)
         plt.show()
 
-    @abstractmethod
-    def get_x_label(self):
-        return "default x label"
-
-    @abstractmethod
-    def get_y_label(self):
-        return "default y label"
-
-    @abstractmethod
-    def columns_to_group_for_plots(self) -> list[str]:
-        """
-        TODO better name
-
-        Generate a plot for each unique tuple of values for the specified columns.
-        Example: ["k", "n"] - a plot will be generated for each unique tuple value (k, n).
-        Example: [] - generate a single plot with the entire data.
-
-        Returns:
-            List of column names to group by.
-        """
-        return []
-
-    @abstractmethod
-    def columns_to_group_in_a_plot(self):
-        """
-        TODO better name
-
-        Generate a data line for each unique value in the specified columns.
-
-        Example: ["impl"] - a line will be generated for each unique `impl` label.
-        """
-
-    @abstractmethod
-    def compute_aggregates(self, data, *, quantum_factor):
-        return data
-
-    @abstractmethod
-    def get_x_axis_column(self) -> str:
-        """
-        TODO better name
-        Column to plot along x axis
-        """
-        return ""
-
-    @abstractmethod
-    def columns_to_plot(self) -> dict[str, tuple[str, str]]:
-        """
-        Dictionary of columns to display in the plot.
-            Key: is the column name in the dataframe
-            Value: Column display name, Marker to use in the plot
-
-        Example:
-            {"c": ("Classical", "o"), "q": ("Quantum", "x")}
-        """
-        return {}
-
-    def color_for_impl(self, impl):
+    def color_for_impl(self, impl: str):
         """
         Returns:
              a color for a given key `impl`, and generates a new unique color if it does not exist.
@@ -162,7 +169,7 @@ class PlottingStrategy(ABC):
         self.colors[impl] = new_color
         return new_color
 
-    def make_plot_title(self, plot_params) -> str:
+    def make_plot_title(self, plot_params: list) -> str:
         columns = [
             f"{column} = {value}"
             for (column, value) in zip(self.columns_to_group_for_plots(), plot_params)
@@ -170,4 +177,4 @@ class PlottingStrategy(ABC):
         return ", ".join(columns)
 
     def serialize_impl_params(self, impl_params: list) -> str:
-        return ", ".join(impl_params)
+        return ", ".join(map(lambda param: f"{param}", impl_params))
